Moves confirm payment intent function into confirm payment intent function
