--- components/ui/checkoutForm.tsx
@@ -1,21 +1,0 @@
-import { PaymentElement } from "@stripe/react-stripe-js";
-
-const CheckoutForm = () => {
-  return (
-    <form>
-      <PaymentElement
-        options={{
-          business: { name: "CourtBooking" },
-          layout: {
-            type: "accordion",
-            defaultCollapsed: false,
-            radios: false,
-            spacedAccordionItems: true,
-          },
-        }}
-      />
-    </form>
-  );
-};
-
-export default CheckoutForm;
